backend/requirements.txt
Django==5.0.3
djangorestframework==3.14.0
django-cors-headers==4.3.1
sentence-transformers==2.7.0
requests==2.31.0
beautifulsoup4==4.12.3
selenium==4.18.1
webdriver-manager==4.0.1
openai==1.14.1
python-dotenv==1.0.1
numpy>=1.24
mysqlclient==2.2.4
cryptography==42.0.5

